Restructure image painting strategy
Switch from drawing all images (components) on a canvas and refreshing that
canvas 50 times per second, to a canvas-less stack of image elements in the
document and no frame refresh. Images (components) are placed before all
"displays" which are placed before all buttons/input fields. This allows
squeezing the cursor image (for tool pickup) between all images/displays
and buttons/input fields so that the cursor image doesn't block the
button/input field but is on top of everything else.